+++ Domain/Entities/Descuento.cs
@@ -4,7 +4,7 @@
     {
         public Guid IdDescuento { get; set; }
         public DateTime FechaInicioVigencia { get; set; }
-        public double Porcentaje { get; set; }
+        public decimal Porcentaje { get; set; }
 
         public ICollection<Recibo> Recibos { get; set; }
     }
